build 74
Corrected error with creating coord.LOtoMRGS and LLtoMGRS internal slmod
functions. Any MGRS related feature that was broken should now work.
Tweaks to stats display. Extra lines should no longer be added. But the
columns are still out of alignment depending on the text
Fixed the Mig-29A appearing as a weapon when used as a kamikazi attack