multicrew stats update
Made changes to account for multicrew stats. Each client in the aircraft shares the stats. However stats get saved based on the seat a player is in.

Need to fully test punish/forgive logic with multicrew.

Started adding admin commands to return the current penalty score of a given user.