Multicrew changes, stats buffer, json, and WIP toward a future feature
-Added option for json stats export.
-- It is a simple boolean setting that will save a copy of any stats file that is written in slmodStats.lua as a json. This includes the main stats file, penalty stats, campaign stats, and mission stats.

- Added stats writting buffer. This will reduce the amount of writes to the stats file. For example instead of updating an entry for 20 rockets one shot event at a time, it will just save it as +20.

Fixed: Dakka in Huey and Mi-8 naming for multicrew seats.
Fixed: Multicrew defs and auto generation of naming if a multicrew unit type is not pre-defined.

Fixed: Crash stat not counting when AI stats are disabled and player is killed by AI.
Added WIP Feature: Best and Current life stats. Setting will be visible in the config file but is not enabled in the code via commented out lines.