moved json conversion
Moved the conversion of a json file to only run if it is enabled. Note there is possibly a bug with converting to json with extremely large files. If you enable it I recommend checking DCS.log to see if slmodTests.lua is loaded. If not then you should disable the json setting and restart.

Also removed some logging from the slmodConfig.